// AdminApp/src/app/pages/SetupPages/RegistrationFormAmend/Model/Step3DTOAmend.ts
interface Step3DtoAmend {
    isWebsite: boolean;
    isSocialMedia: boolean;
    isMobileApplication: boolean;
    websiteUrls: string;

    socialMediaUrls: string;
    influencer: string;
    mobileApplicationUrl: string;
    companyOrIndividualType: string;

    isWebsiteAmend: boolean;
    isSocialMediaAmend: boolean;
    isMobileApplicationAmend: boolean;
}

export default Step3DtoAmend;